fantasy/fantasy/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10,<=3.12"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "google-genai (>=1.38.0,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "google-adk (==1.17.0)",
    "rapidfuzz (>=3.14.1,<4.0.0)",
    "sortedcontainers (>=2.4.0,<3.0.0)",
    "aiolimiter (>=1.2.1,<2.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "aiohttp (>=3.13.0,<4.0.0)",
]

[tool.poetry]
packages = [{include = "agents", from = "src"}]

[tool.poetry.group.local.dependencies]
common = { path = "../common", develop = true }

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
